+++ iLynx.UI/Controls/IControl.cs
@@ -4,8 +4,8 @@ namespace iLynx.UI.Controls
 {
     public interface IControl : IUIElement
     {
-        float Width { get; set; }
-        float Height { get; set; }
+        uint Width { get; set; }
+        uint Height { get; set; }
         Point Position { get; set; }
         Color Background { get; set; }
         Color Foreground { get; set; }
